Change logic of getting random question to fetch from own db
